feat: add sent message detail screen with delete functionality and UI components

+++ data/remote/room/dto/RoomResponse.kt
@@ -31,7 +31,7 @@ data class SentMessageResponse(
     val initialMessage: String,
     val confessionTitle: String,
     val confessionMessage: String,
-    val channelMessageId: Long,
+    val channelMessageId: Int,
     val createdAt: String,
     val status: String
 )
